Typo in the content
While reading the tutorial, I found that this sentence: "if blah.c is modified, running make should recompile the file. And conversely, if blah.c has not changed, then it should be recompiled." seems incorrect. The second part should be negated and be "if blah.c has not changed, then it should not be recompiled.".