Change confusing "assign"s to "set"s
The form "assign a variable to a value" was used multiple times in the document. This is confusing; it sounds like the variable was put into the value. You could say "assign a value to a variable", but to keep the format consistent I changed these to "set a variable to a value".